added default and authorization error components
